+++ packages/diagram/src/overlays/types.ts
@@ -1,38 +1,34 @@
-import type { ActorRefFrom } from 'xstate'
 import type {
+  ElementDetailsActorRef,
+  ElementDetailsInput,
   ElementDetailsLogic,
-  Events as ElementDetailsEvents,
-  Input as ElementDetailsInput,
 } from './element-details/actor'
 import type {
-  Events as RelationshipDetailsEvents,
-  Input as RelationshipDetailsInput,
+  RelationshipDetailsActorRef,
+  RelationshipDetailsInput,
   RelationshipDetailsLogic,
 } from './relationship-details/actor'
 import type {
-  Events as RelationshipsBrowserEvents,
-  Input as RelationshipsBrowserInput,
+  RelationshipsBrowserActorRef,
+  RelationshipsBrowserInput,
   RelationshipsBrowserLogic,
 } from './relationships-browser/actor'
 export namespace Overlays {
   export namespace ElementDetails {
-    export type Events = ElementDetailsEvents
-    export type Input = ElementDetailsInput
-    export type Logic = ElementDetailsLogic
-    export interface ActorRef extends ActorRefFrom<ElementDetailsLogic> {}
+    export interface Input extends ElementDetailsInput {}
+    export interface Logic extends ElementDetailsLogic {}
+    export interface ActorRef extends ElementDetailsActorRef {}
   }
 
   export namespace RelationshipDetails {
-    export type Events = RelationshipDetailsEvents
     export type Input = RelationshipDetailsInput
     export interface Logic extends RelationshipDetailsLogic {}
-    export interface ActorRef extends ActorRefFrom<RelationshipDetailsLogic> {}
+    export interface ActorRef extends RelationshipDetailsActorRef {}
   }
 
   export namespace RelationshipsBrowser {
-    export type Events = RelationshipsBrowserEvents
-    export type Input = RelationshipsBrowserInput
+    export interface Input extends RelationshipsBrowserInput {}
     export interface Logic extends RelationshipsBrowserLogic {}
-    export interface ActorRef extends ActorRefFrom<RelationshipsBrowserLogic> {}
+    export interface ActorRef extends RelationshipsBrowserActorRef {}
   }
 }